most: fix race between enqueue and most_stop_enqueue
The "broken in pipe" handler of the USB-HDM calls most_stop_enqueue() to
stop the MBO traffic prior to return all MBOs back to the Mostcore.  As
the enqueue() call from the Mostcore may run in parallel with the
most_stop_enqueue(), the HDM may run into the inconsistent state and
crash the kernel.

This patch wraps enqueue(), most_stop_enqueue() and
most_resume_enqueue() by a mutex to remove the race between them.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>