staging: most: do not return error from most_submit_mbo
There is no way to recover runtime from the case where the function
most_submit_mbo() returns the error code, because it happens only if the
mbo is NULL or mbo->context is manipulated by the calling module what
are the signs of the wrong code integration.

This patch replaces the return type of the function most_submit_mbo()
with 'void' to prevent even the tries to silently solve the integration
errors by the runtime code like call the function most_put_mbo() or
change the state of the module.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>