staging: most: do not call most_put_mbo if most_submit_mbo fails
The function most_submit_mbo() returns the error only if the mbo is NULL
or mbo->context is NULL.  The function most_put_mbo() uses mbo and
mbo->context without any checks.  That all means that the calling the
function most_put_mbo() in case where the function most_put_mbo() fails
is useless and faulty.

This patch removes call of the most_put_mbo() in case where the
most_submit_mbo() fails.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>