most: cdev-aim: make write for streaming channels user friendly
This patch allows to call the write() function for the synchronous and
isochronous channels with any size of the user buffer.  The AIM waits
until the MBO buffer of the streaming channel is full before it is
forwarded to the HDM.

The new behavior is backward compatible to the old applications, since
all known applications needed to fill the buffer completely anyway.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>