staging: most: round up dbr memory for async/ctrl
As the DBR memory is allocated as a number of units of a size
DBR_BLOCK_SIZE, it is more efficient to configure the DIM2 IP to use
whole blocks of the DBR memory.  This patch rounds up the sizes of the
DBR memory used by the DIM2 IP for async/ctrl channels.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>