staging: most: workaround: monitor atx dbr space
This patch implements workaround for a DIM2 issue where the DIM2 macro,
in the case where the asynchronous channel sends the data (to MOST),
sporadically duplicates the synchronous blocks in a size of the half
synchronous DBR buffer.

The patch monitors the size of the unused asynchronous Tx DBR memory
(that includes the ping and pong sizes) to prevent the potential DBR
overflow for the asynchronous Tx DBR buffer.

The patched DIM2 HDM expects that the platform driver delivers the 2nd
platform irq (index 1) as the mlb_int of the DIM2 macro.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>