staging: most: fix networking operation for some startup cases
In the case where the networking interface based on the MediaLB starts
before the MOST ring, the networking interface remains disabled even
after the MOST ring is active.

This patch removes the dependency on the MOST link status and keeps the
networking queue active to keep the networking interface working for the
described case.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>