staging: most: setup mac address before ifup is finished
Current networking AIM implements non-standard behavior where it starts
the networking interface with an invalid MAC address and changes it by
the availability of the valid one.

This patch introduces the completion to set the MAC address of the
networking interface before the .ndo_open callback (ifup) of the
net_device is finished.  Additionally, it tracks different problems
around the MAC setup.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>