staging: most: hdm-usb: remove filtering of networking state
In the case where the networking interface goes down and then up, the
USB HDM does not report the INIC state even after the request from the
networking AIM, what may lead to non-functional network.

Additionally, the networking AIM does the full cycle of the MAC checking
and tracing.

This patch removes impractical filtering of the MAC address and link
status events.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>