staging: most: remove dependency on networking-aim
The usb-hdm as well as the dim2-hdm depend on the networking-aim because
they use the function most_deliver_netinfo exported from the
networking-aim.

To remove this dependency, this patch replaces the call of the function
most_deliver_netinfo with the call of the function that is the parameter
'on_netinfo' of the function request_netinfo.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>