most: move autoconf feature from source to patch
This moves the autoconf feature from the source code to the patch
mld/patches/feature__core_autoconf.patch and moves the
drivers/staging/most/mostcore/default_conf.c as an example of the configuration
for the autoconf to the mld/default_conf.c.

Additionally, this adapts the script fetch-mld.bash and the Makefile.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>