staging: most: usb: clear functional stall on OUT endpoint
This backports the
commit 8f20f2dca8b390b4fe9e40ed6196c805a3f07b66
Author: Christian Gromm <christian.gromm@microchip.com>
Date:   Tue Nov 21 15:05:15 2017 +0100

For the MOST packet channel there are two dedicated USB endpoints.  But
internally the hardware has actually one channel for data forwarding
from and to MOST.  To have the hardware clean up its state machine
correctly in case of an error, both USB pipes need to be reset.  This
patch triggers the host to also clear the OUT endpoint's halt condition
in case an IN endpoint has signaled to be stalled.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@microchip.com>