staging: most: let I2S HDM do the coherent memory allocation
On arm64/aarch64 architectures the allocation of coherent memory
needs a device that has the dma_ops proberly set. That's why the
core module of the MOST driver is no longer able to allocate this
kind of memory. This patch moves the allocation process down to
the I2S HDM layer where the proper devices exist in either the
USB host controller or the platform device.