staging: most: usb: change synchronization lock order
The function most_stop_enqueue() locks the mutex nq_mutex.
The thread calling the callback hdm_enqueue() locks the same mutex.
The callback hdm_enqueue() locks the mutex io_mutex.

Following mutex deadlock is possible if the most_stop_enqueue() is
called within the locked io_mutex:
most_stop_enqueue() waits for the nq_mutex, that is indirectly locked by
the hdm_enqueue() that is waiting for the io_mutex.

The same condition is possible for the most_register_interface().

This patch moves the most_stop_enqueue() and most_resume_enqueue() out
of mutex io_mutex.

Signed-off-by: Andrey Shvetsov <andrey.shvetsov@k2l.de>
Signed-off-by: Christian Gromm <christian.gromm@microchip.com>