staging: most: usb: make assignment to pointer after it has been declared
This patch splits the declaration of the MOST device pointer and
its assignment. It is needed to have sanity checks on interface
pointers before they are being used in a container_of construction
to retrieve the value of the MOST device pointer. To have a common
pattern this is applied to other assignments as well.

Signed-off-by: Christian Gromm <christian.gromm@microchip.com>